Refactor database structure
