+++ drizzle.config.ts
@@ -2,5 +2,5 @@ import { defineConfig } from 'drizzle-kit'
 
 export default defineConfig({
   dialect: 'postgresql',
-  schema: './src/db/schema.ts',
+  schema: './src/database/schema.ts',
 })
